Fixes for #648 and #718
PowerShell/Win32-OpenSSH#648
PowerShell/Win32-OpenSSH#718

+++ contrib/win32/openssh/OpenSSHBuildHelper.psm1
@@ -339,7 +339,7 @@ function Package-OpenSSH
 
     if ($DestinationPath -ne "") {
         if (Test-Path $DestinationPath) {            
-            Remove-Item $DestinationPath\* -Force -Recurse
+            Remove-Item $DestinationPath\* -Force -Recurse -ErrorAction SilentlyContinue
         }
         else {
             New-Item -ItemType Directory $DestinationPath -Force | Out-Null
@@ -542,18 +542,22 @@ function UnInstall-OpenSSH
         [string]$OpenSSHDir = "$env:SystemDrive\OpenSSH"
     )
 
+    if (-not (Test-Path $OpenSSHDir))
+    {
+        return
+    }
+
     Push-Location $OpenSSHDir
     if((Get-Service ssh-agent -ErrorAction Ignore) -ne $null) {
         Stop-Service ssh-agent -Force
     }
     &( "$OpenSSHDir\uninstall-sshd.ps1")
-    &( "$OpenSSHDir\uninstall-sshlsa.ps1")
         
     $machinePath = [Environment]::GetEnvironmentVariable('Path', 'MACHINE')
     $newMachineEnvironmentPath = $machinePath
     if ($machinePath.ToLower().Contains($OpenSSHDir.ToLower()))
     {        
-        $newMachineEnvironmentPath.Replace("$OpenSSHDir;", '')
+        $newMachineEnvironmentPath = $newMachineEnvironmentPath.Replace("$OpenSSHDir;", '')
         $env:Path = $env:Path.Replace("$OpenSSHDir;", '')
     }
